+++ odoo/local-src/stock_lot_loss/tests/test_lot_loss.py
@@ -21,7 +21,26 @@ class TestLotLoss(SavepointCase):
                 'tracking': 'lot',
             }
         )
-
+        cls.product_2 = cls.env['product.product'].create(
+            {
+                'name': 'Product 2 stock_lot_loss no tracking',
+                'type': 'product',
+                'uom_id': cls.env.ref('product.product_uom_unit').id,
+                'uom_po_id': cls.env.ref('product.product_uom_unit').id,
+                'default_code': 'Code product lot_loss no tracking 2',
+                'tracking': 'none',
+            }
+        )
+        cls.product_3 = cls.env['product.product'].create(
+            {
+                'name': 'Product 3 stock_lot_loss no tracking',
+                'type': 'product',
+                'uom_id': cls.env.ref('product.product_uom_unit').id,
+                'uom_po_id': cls.env.ref('product.product_uom_unit').id,
+                'default_code': 'Code product lot_loss no tracking 3',
+                'tracking': 'none',
+            }
+        )
         wh = cls.env['stock.warehouse'].search([])
         cls.location = wh[0].view_location_id
         cls.location.usage = 'internal'
@@ -111,6 +130,78 @@ class TestLotLoss(SavepointCase):
         )
         self.assertEqual(len(quants), 2)
 
+    def initiate_values_no_tracking(self):
+        # Create picking 2
+        self.picking_2 = self.env['stock.picking'].create(
+            {
+                'picking_type_id': self.pick_type.id,
+                'location_id': self.location.id,
+                'location_dest_id': self.loc_customer.id,
+            }
+        )
+        self.move_2 = self.env['stock.move'].create(
+            {
+                'picking_id': self.picking_2.id,
+                'name': 'Test move 2',
+                'product_id': self.product_2.id,
+                'product_uom': self.product_2.uom_id.id,
+                'product_uom_qty': 6,
+                'location_id': self.location.id,
+                'location_dest_id': self.loc_customer.id,
+                'date': '2018-01-01 00:00:00',
+            }
+        )
+        self.move_2.action_confirm()
+        self.move_3 = self.env['stock.move'].create(
+            {
+                'picking_id': self.picking_2.id,
+                'name': 'Test move 3',
+                'product_id': self.product_3.id,
+                'product_uom': self.product_3.uom_id.id,
+                'product_uom_qty': 1,
+                'location_id': self.location.id,
+                'location_dest_id': self.loc_customer.id,
+                'date': '2018-01-01 00:00:00',
+            }
+        )
+        self.move_3.action_confirm()
+
+        # Put product in stock
+        # Product2: 3
+        # Product3: 5
+        inventory = self.env['stock.inventory'].create(
+            {
+                'name': 'Test',
+                'filter': 'product',
+                'location_id': self.location.id,
+                'product_id': self.product_2.id,
+            }
+        )
+        inventory.prepare_inventory()
+        inventory.line_ids.unlink()
+        inventory.line_ids.create(
+            {
+                'product_id': self.product_2.id,
+                'product_qty': 3.0,
+                'inventory_id': inventory.id,
+                'location_id': self.location.id,
+            }
+        )
+        inventory.line_ids.create(
+            {
+                'product_id': self.product_3.id,
+                'product_qty': 5.0,
+                'inventory_id': inventory.id,
+                'location_id': self.location.id,
+            }
+        )
+        inventory.action_done()
+
+        quants = self.env['stock.quant'].search(
+            [('product_id', 'in', (self.product_2.id, self.product_3.id))]
+        )
+        self.assertEqual(len(quants), 2)
+
     def test_lot_loss_line1(self):
         """ Create loss of line1 """
         # There should be 2 quant in stock
@@ -132,7 +223,9 @@ class TestLotLoss(SavepointCase):
         pack_lot_A.qty = 1
         op.save()
 
-        pack_lot_A.with_context(round_autoset=False)._skip_lot()
+        op.with_context(round_autoset=False)._skip_operation(
+            pack_op_lot_id=pack_lot_A
+        )
 
         # Check new pack operation
         new_op = self.picking_1.pack_operation_ids
@@ -240,7 +333,9 @@ class TestLotLoss(SavepointCase):
         pack_lot_B.qty = 1
         op.save()
 
-        pack_lot_B.with_context(round_autoset=False)._skip_lot()
+        op.with_context(round_autoset=False)._skip_operation(
+            pack_op_lot_id=pack_lot_B
+        )
 
         # Check new pack operation
         new_op = self.picking_1.pack_operation_ids
@@ -319,6 +414,94 @@ class TestLotLoss(SavepointCase):
                 'inventory_id': inventory.id,
                 'location_id': self.location.id,
                 'prod_lot_id': self.product_1_lotB.id,
+            }
+        )
+        inventory.action_done()
+        self.assertEqual(block_move.state, 'cancel')
+
+    def test_loss_line_no_tracking(self):
+        """ Create loss of product_2 without tracking"""
+        self.initiate_values_no_tracking()
+
+        self.picking_2.with_context(round_autoset=False).action_assign()
+
+        quant = self.picking_2.mapped('move_lines.reserved_quant_ids')
+        self.assertEqual(len(quant), 2)
+
+        ops = self.picking_2.pack_operation_ids
+        self.assertEqual(len(ops), 2)
+
+        op_2 = ops.filtered(lambda op: op.product_id == self.product_2)
+        op_3 = ops.filtered(lambda op: op.product_id == self.product_3)
+        op_2.qty_done = 1.0
+        op_3.qty_done = 1.0
+
+        op_2.with_context(round_autoset=False)._skip_operation()
+        # Check new pack operation
+        new_op_2 = self.picking_2.pack_operation_ids.filtered(
+            lambda op: op.product_id == self.product_2
+        )
+        self.assertNotEqual(op_2, new_op_2)
+        self.assertEqual(new_op_2.product_qty, 1)
+        self.assertEqual(new_op_2.qty_done, 1)
+
+        # Check blocking move has been created
+        loss_picking_type = self.env.ref(
+            'stock_lot_loss.stock_picking_type_23'
+        )
+        block_move = (
+            self.env['stock.quant']
+            .search(
+                [
+                    ('qty', '>', 0.0),
+                    ('product_id', '=', self.product_3.id),
+                    ('location_id', '=', self.location.id),
+                    (
+                        'reservation_id.picking_id.picking_type_id',
+                        '=',
+                        loss_picking_type.id,
+                    ),
+                ]
+            )
+            .mapped('reservation_id')
+        )
+        self.assertEqual(block_move.ids, [])
+        block_move = (
+            self.env['stock.quant']
+            .search(
+                [
+                    ('qty', '>', 0.0),
+                    ('product_id', '=', self.product_2.id),
+                    ('location_id', '=', self.location.id),
+                    (
+                        'reservation_id.picking_id.picking_type_id',
+                        '=',
+                        loss_picking_type.id,
+                    ),
+                ]
+            )
+            .mapped('reservation_id')
+        )
+        self.assertEqual(block_move.state, 'assigned')
+        self.assertEqual(block_move.product_qty, 2)
+
+        # Check blocked lot cleanup
+        inventory = self.env['stock.inventory'].create(
+            {
+                'name': 'Test',
+                'filter': 'product',
+                'location_id': self.location.id,
+                'product_id': self.product_1.id,
+            }
+        )
+        inventory.prepare_inventory()
+        inventory.line_ids.unlink()
+        inventory.line_ids.create(
+            {
+                'product_id': self.product_2.id,
+                'product_qty': 1.0,
+                'inventory_id': inventory.id,
+                'location_id': self.location.id,
             }
         )
         inventory.action_done()