+++ odoo/local-src/stock_reassign_auto/__manifest__.py
@@ -10,6 +10,7 @@
     'depends': [
         'stock_picking_assignment',
         'stock_picking_subcode',
+        'queue_job',
     ],
     'data': [
     ],